separate out createProject factory as a module for easier maintainability of code
